back/requirements.txt
django==3.1.7
requests==2.25.1
djangorestframework==3.12.2
drf-yasg==1.20.0
django-cors-headers==3.7.0
djangorestframework-simplejwt==4.6.0
mysqlclient==2.0.3
django-mysql==3.8.1
gunicorn